ES1-1250,ES1-1224 : Live no progress events , CDVR recordings error
Reason for change:

Realtek will provide the position in absolute format when the qtdemux override (PTS restamping) is disabled.
However, if the first PTS value isn't sent to the pipeline, the segment event query will return 0.
Consequently, the player won't be able to calculate the position during tuning when the first PTS is non-zero.
The following code block aids in obtaining the correct segment start value for the scenario where QTDEMUX is disabled, specifically for ES1 and in the mac simulator.

Test Procedure: Test case updated in the ticket
Risks: low


ES1-1250,ES1-1224 : Live no progress events , CDVR recordings error

Reason for change:

Followup change to avoid test build failure

Test Procedure: Test case updated in the ticket
Risks: low

(cherry picked from commit 8642139c35427a8cafb05cb6e6ef1da8180a3c30)

+++ AampStreamSinkInactive.h
@@ -142,7 +142,7 @@ public:
      *   @fn GetPositionMilliseconds
 	 *   @brief stub implementation for Inactive aamp instance
 	 */
-	virtual long GetPositionMilliseconds(void)
+	virtual long long GetPositionMilliseconds(void)
 	{
 		AAMPLOG_WARN("Called AAMPGstPlayer()::%s stub", __FUNCTION__);
 		return 0;
